[#46] [fix] Resolve the console error with dynamic rendering
